requirements.txt
Django==5.0.7
djangorestframework==3.15.2
mysqlclient==2.2.4
pytz